Add a very dummy high-pass filter
I have coded the function `filter_noise`, that it is located in
`filter.jl`. It is a very simple high-pass filter that it is coded
in order to eliminate the very low frequencies spurious signal,
in particular, the one-over-f noise. It does not work for the
atmospheric noise.

Now the strip2 module, can create random sky representation based
on a given CMB power spectrum.

I have been start coding a map-maker. In order to resolve the
map-maker inverse problem, for TODs comphensive of noise sources,
I have developed the ConjGrad.jl Julia package, that is now
available on my personal github page.

As I thought, the map-makring is particulary tricky to code.
If I have time i will deploy the whole package, comphensive of
map-maker, in a few days.

+++ src/noise_map.jl
@@ -1,4 +1,4 @@
-function white_noise(N, pix_size, wnl)
+function white_noise(N::Int, pix_size::Float64, wnl::Float64)
 
     rng = MersenneTwister()
     white_noise_map = randn(rng, N, N) .* (wnl / pix_size)
@@ -8,7 +8,7 @@ function white_noise(N, pix_size, wnl)
 end
 
 
-function atmospheric_noise(N, pix_size, anl)
+function atmospheric_noise(N::Int, pix_size::Float64, anl::Float64)
 
     rng = MersenneTwister()
     random_realization = randn(rng, N, N)
@@ -28,7 +28,7 @@ function atmospheric_noise(N, pix_size, anl)
 end
 
 
-function one_over_f(N, pix_size, oofnl)
+function one_over_f(N::Int, pix_size::Float64, oofnl::Float64)
 
     rng = MersenneTwister()
     random_realization = randn(rng, N, N)
